# docs/openapi-v1.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "BriskDB HTTP API",
    "version": "1",
    "description": "Versioned BriskDB data and administration machine API. The API version is independent of the Cargo package and storage-format versions.",
    "license": {
      "name": "MIT"
    }
  },
...
  "paths": {
    "/v1": {
      "get": {
        "tags": [
          "data"
        ],
        "operationId": "getApiVersion",
        "responses": {
          "200": {
            "description": "HTTP API version",
            "headers": {
              "BriskDB-API-Version": {
                "$ref": "#/components/headers/BriskDBApiVersion"
              },
              "BriskDB-Request-ID": {
                "$ref": "#/components/headers/BriskDBRequestId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiVersion"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem400"
          },
          "501": {
            "$ref": "#/components/responses/Problem501"
          }
        },
        "summary": "Discover HTTP API version",
        "servers": [
          {
            "url": "http://127.0.0.1:7654",
            "description": "Data listener (default address)"
          }
        ],
        "x-briskdb-plane": "data",
        "parameters": [
          {
            "$ref": "#/components/parameters/BriskDBRequestId"
          }
        ]
      },
      "head": {
        "tags": [
          "data"
        ],
        "operationId": "headApiVersion",
        "responses": {
          "200": {
            "description": "HTTP API version",
            "headers": {
              "BriskDB-API-Version": {
                "$ref": "#/components/headers/BriskDBApiVersion"
              },
              "BriskDB-Request-ID": {
                "$ref": "#/components/headers/BriskDBRequestId"
              },
              "Content-Length": {
                "$ref": "#/components/headers/ContentLength"
              },
              "Content-Type": {
                "description": "Media type of the corresponding GET representation.",
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "application/json"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or argument",
            "headers": {
              "BriskDB-API-Version": {
                "$ref": "#/components/headers/BriskDBApiVersion"
              },
              "BriskDB-Request-ID": {
                "$ref": "#/components/headers/BriskDBRequestId"
              },
              "Content-Length": {
                "$ref": "#/components/headers/ContentLength"
              },
              "Content-Type": {
                "description": "Media type of the corresponding GET representation.",
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "application/problem+json"
                }
              }
            }
          },
          "501": {
            "description": "Unsupported operation or globally rejected request control",
            "headers": {
              "BriskDB-API-Version": {
                "$ref": "#/components/headers/BriskDBApiVersion"
              },
              "BriskDB-Request-ID": {
                "$ref": "#/components/headers/BriskDBRequestId"
              },
              "Content-Length": {
                "$ref": "#/components/headers/ContentLength"
              },
              "Content-Type": {
                "description": "Media type of the corresponding GET representation.",
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "application/problem+json"
                }
              }
            }
          }
        },
        "summary": "Inspect headers for discover http api version",
        "servers": [
          {
            "url": "http://127.0.0.1:7654",
            "description": "Data listener (default address)"
          }
        ],
        "x-briskdb-plane": "data",
        "parameters": [
          {
            "$ref": "#/components/parameters/BriskDBRequestId"
          }
        ],
        "x-briskdb-implicit-head": true
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiVersion": {
        "type": "object",
        "properties": {
          "api_version": {
            "type": "string",
            "const": "1"
          },
          "value_encoding": {
            "type": "string",
            "const": "legacy-json-v1"
          },
          "supported_value_encodings": {
            "type": "array",
            "prefixItems": [
              {
                "type": "string",
                "const": "legacy-json-v1"
              },
              {
                "type": "string",
                "const": "lossless-json-v1"
              }
            ],
            "items": false,
            "minItems": 2,
            "maxItems": 2
          },
          "session_scope": {
            "type": "string",
            "const": "request"
          },
          "sql_dialect": {
            "type": "string",
            "const": "sqlite"
          },
          "max_request_bytes": {
            "type": "integer",
            "const": 2097152
          },
          "max_result_rows": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000000
          },
          "max_result_logical_bytes": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1073741824
          },
          "stream_buffer_rows": {
            "type": "integer",
            "const": 16
          },
          "request_id_header": {
            "type": "string",
            "const": "BriskDB-Request-ID"
          },
          "idempotency_key_header": {
            "type": "string",
            "const": "BriskDB-Idempotency-Key"
          },
          "stream_media_type": {
            "type": "string",
            "const": "application/x-ndjson; charset=utf-8"
          }
        },
        "required": [
          "api_version",
          "value_encoding",
          "supported_value_encodings",
          "session_scope",
          "sql_dialect",
          "max_request_bytes",
          "max_result_rows",
          "max_result_logical_bytes",
          "stream_buffer_rows",
          "request_id_header",
          "idempotency_key_header",
          "stream_media_type"
        ]
      },
...
      "OpaqueId": {
        "allOf": [
          {
            "type": "string",
            "pattern": "^[0-9a-f]{32}$"
          },
          {
            "not": {
              "const": "00000000000000000000000000000000"
            }
          }
        ],
        "examples": [
          "0123456789abcdef0123456789abcdef"
        ]
      },
...
      "ProblemDetails400": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "const": 400
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ]
      },
...
      "ProblemDetails501": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "const": 501
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ]
      },
...
    },
    "parameters": {
      "BriskDBRequestId": {
        "name": "BriskDB-Request-ID",
        "in": "header",
        "required": false,
        "description": "Optional caller-supplied correlation ID. It must occur exactly once and use the canonical nonzero 128-bit lowercase hexadecimal form.",
        "schema": {
          "$ref": "#/components/schemas/OpaqueId"
        }
      },
...
    },
    "headers": {
      "BriskDBApiVersion": {
        "description": "Version of the HTTP representation contract.",
        "required": true,
        "schema": {
          "type": "string",
          "const": "1"
        }
      },
      "BriskDBRequestId": {
        "description": "Canonical request correlation ID supplied or generated for this attempt.",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/OpaqueId"
        }
      },
...
      "ContentLength": {
        "description": "The size of the corresponding GET representation; a HEAD response carries no body.",
        "required": true,
        "schema": {
          "type": "integer",
          "minimum": 0,
          "maximum": 18446744073709551615
        }
      },
...
    },
    "responses": {
      "Problem400": {
        "description": "Invalid request or argument",
        "headers": {
          "BriskDB-API-Version": {
            "$ref": "#/components/headers/BriskDBApiVersion"
          },
          "BriskDB-Request-ID": {
            "$ref": "#/components/headers/BriskDBRequestId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails400"
            }
          }
        }
      },
...
      "Problem501": {
        "description": "Unsupported operation or globally rejected request control",
        "headers": {
          "BriskDB-API-Version": {
            "$ref": "#/components/headers/BriskDBApiVersion"
          },
          "BriskDB-Request-ID": {
            "$ref": "#/components/headers/BriskDBRequestId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails501"
            }
          }
        }
      },
...
    }
  },
...
}